gave up on setuptools post-install
Package is installed to a temporary build directory first, and post
install scripts are run there, so we can't influence the user's
project directory or import modules from virtualenv packages.

And making post-install message doesn't look pretty either, it is
lost in the midst of build messages of pip.

So user is expected to read the Readme.md file first and follow the
instructions there, like downloading waf, copying wscript, etc.